feat: 增加排序核心代码测试用例
