+++ packages/table/src/utils/index.ts
@@ -13,6 +13,7 @@ import type {
   ProColumnType,
   UseFetchDataAction,
 } from '../typing';
+import arrayMove from 'array-move';
 
 /**
  * 检查值是否存在 为了 避开 0 和 false
@@ -199,4 +200,21 @@ export function parseDefaultColumnConfig<T, Value>(columns: ProColumns<T, Value>
     }
   });
   return { sort, filter };
+}
+
+export type SortDataParams = { oldIndex: number; newIndex: number };
+
+/**
+ * 数据排序核心逻辑
+ *
+ * @param oldIndex 原始位置
+ * @param newIndex 新位置
+ * @param data 原始数组
+ */
+export function sortData<T>({ oldIndex, newIndex }: SortDataParams, data: T[]): T[] | null {
+  if (oldIndex !== newIndex) {
+    const newData = arrayMove([...(data || [])], oldIndex, newIndex).filter((el) => !!el);
+    return [...newData];
+  }
+  return null;
 }
