+++ tests/table/column.test.tsx
@@ -138,74 +138,4 @@ describe('Table ColumnSetting', () => {
     await waitForComponentToPaint(html, 1200);
     expect(html.find('td.ant-table-cell')).toMatchSnapshot();
   });
-  it('🎏 [dragSort] render drag sort default handle by dragSortKey', async () => {
-    const html = mount(
-      <ProTable
-        size="small"
-        columns={[
-          {
-            title: '排序',
-            dataIndex: 'sort',
-          },
-          {
-            title: '姓名',
-            dataIndex: 'name',
-            className: 'drag-visible',
-          },
-          {
-            title: '年龄',
-            dataIndex: 'age',
-          },
-          {
-            title: '地址',
-            dataIndex: 'address',
-          },
-        ]}
-        request={request}
-        rowKey="key"
-        dragSortKey="sort"
-      />,
-    );
-    await waitForComponentToPaint(html, 1200);
-    expect(html.find('.dragSortDefaultHandle')).toMatchSnapshot();
-  });
-
-  it('🎏 [dragSort] render drag sort custom handle by dragSortHandlerRender', async () => {
-    const dragHandleRender = (rowData: any, idx: any) => (
-      <>
-        <MenuOutlined className="dragSortCustomHandle" style={{ cursor: 'grab', color: 'gold' }} />
-        {idx + 1} - {rowData.name}
-      </>
-    );
-    const html = mount(
-      <ProTable
-        size="small"
-        columns={[
-          {
-            title: '排序',
-            dataIndex: 'sort',
-          },
-          {
-            title: '姓名',
-            dataIndex: 'name',
-            className: 'drag-visible',
-          },
-          {
-            title: '年龄',
-            dataIndex: 'age',
-          },
-          {
-            title: '地址',
-            dataIndex: 'address',
-          },
-        ]}
-        request={request}
-        rowKey="key"
-        dragSortKey="sort"
-        dragSortHandlerRender={dragHandleRender}
-      />,
-    );
-    await waitForComponentToPaint(html, 1200);
-    expect(html.find('.dragSortCustomHandle')).toMatchSnapshot();
-  });
 });
